+++ src/components/screen/screen-form.scss
@@ -1,4 +1,4 @@
 .resolution-plus-icon {
-  margin-top: 50px;
+  margin: 45px 6px 0 6px;
   padding: 0;
 }
